Revamp GitHub workflows and rename blank.yml to github-snake.yml
This commit modifies the update-readme.yml script for better automation, adds a new github-metrics.yml file for a compact display of GitHub metrics and renames blank.yml to github-snake.yml for clarity and relevance. These changes improve the overall workflow execution and enhance the readability and organization of the repository.