Update image source for Yandex Cloud icon
The source for the Yandex Cloud icon image in the README has been changed from a lengthy external URL to a locally saved .svg file. This change makes the file more manageable, improves load times, and prevents possible broken links in the future if the external source is removed or modified.