Update SQLAlchemy icon source in README.md
Changed the source url for the SQLAlchemy icon in the README.md file. The previous link was broken or not providing the correct icon which resulted in an unresolved image. Update was necessary to ensure correct rendering of all resources and provide a better user experience when interacting with the repository's README.md on Github.